Implemented 2nd order prolongation and test is extended to it. There was a bug in the way the send and recv buffer sizes were calculated for the 2nd order prolongation. This is fixed now. The default is still first order prolongation if nProlongOrderIn is missing.
Test had to be modified to 3x3x3 root with the middle node refined,
otherwise the periodic boundary does not work out for the
prolonged coordinates.

The increased root grid revealed a bug in the original code:
subface was widened even in non-amr directions
(to include extra faces and edges). Now this is also fixed.